Allow to define SQL number format in common/configuration.ini
Currently, all numbers displayed directly through SQL queries (exports and
dashboards) are in the default format:

 - en_US if not set otherwise by the SQL server
 - 2 decimals

This patchs allows to define those preferences in common/configuration.ini,
as such:

[database]
sql_locale = "fr_FR"
sql_decimals = 1

If sql_locale is omitted, it will be the SQL server default (en_US by default)
If sql_decimals is omitted, it will be defaulted to 2.

Please note that is PR is a partial solution to Issue 311: Numbers internationalization.