Update index.php
Greatly reorganizes the install/index file in order to help make it a bit more readable. Still relies on session variables and is more of a functional program style, but at least uses more descriptive variable names and is structured to help the user follow what is happening. Hopefully this will make the advancement into more of an OOP installer a bit easier.