Removes vestigial commands in module installer files
Removes vestigial commands in the module installer files (most often "have_read_write_access_to_config" being a dependency for an upgrade that doesn't actually write to the config file.

Also fixes a situation in the 3.1.0 upgrade of the Licensing module where the configuration file was being rewritten for no reason.